pyproject.toml
[project]
name = "s3_document_processor"
version = "0.1.0"
description = "Process documents from S3/MinIO and convert them to markdown"
requires-python = ">=3.8"
dependencies = [
    "fastapi",
    "uvicorn",
    "python-multipart",
    "boto3",
    "sqlalchemy",
    "python-dotenv",
    "ulid-py",
    "aiofiles",
    "requests",
    "httpx"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
select = ["E", "F", "B", "I"]
